pyproject.toml
[tool.poetry]
name = "mle-aws"
version = "0.1.0"
description = ""
authors = ["Arora, Apoorva <apoorva.arora@kpn.com>"]
readme = "README.md"
packages = [{include = "mle_aws"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25.0"
boto3 = "^1.28.1"
pandas = "^2.0.3"
xgboost = "^1.7.6"
scikit-learn = "^1.3.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
flask = "^2.3.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
